requirements.txt
pygame
metayaml
pillow
noise